feat: add product form component
